Relocate ghidra java scripts, refactor testing classes
