+++ .gitignore
@@ -1,6 +1,8 @@
 # Ignore generated Ghidra projects
 riscv64/exemplars.gpr
 riscv64/exemplars.rep
+riscv64/exemplars.lock
+riscv64/exemplars.lock~
 # Ignore imported and extracted binaries
 riscv64/kernel/System.map*
 riscv64/kernel/vmlinux*
